requirements.txt
absl-py==0.9.0
aiohttp==3.7.3
altgraph==0.17
appdirs==1.4.3
appnope==0.1.0
astor==0.8.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
Automat==20.2.0
awscli==1.18.66
backcall==0.1.0
bcrypt==3.2.0
beautifulsoup4==4.4.1
bleach==3.1.0
blis==0.4.1
boto3==1.13.16
botocore==1.16.16
bs4==0.0.1
cachetools==4.0.0
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.0
defusedxml==0.6.0
demjson==2.2.4
distlib==0.3.0
docutils==0.15.2
docx2txt==0.8
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
filelock==3.0.12
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
html-table-parser==0.1.0
html-table-parser-python3==0.1.5
hyperlink==19.0.0
idna==2.9
imageio==2.6.1
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.1
json2xlsx==1.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-metrics==1.1.0
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.5.0
macholib==1.14
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
modulegraph==0.18
multidict==5.1.0
murmurhash==1.0.2
nameparser==1.0.6
nbconvert==5.6.0
nbformat==4.4.0
netaddr==0.8.0
networkx==2.4
nltk==3.4.5
notebook==6.0.1
numpy==1.17.1
oauthlib==3.1.0
opencv-contrib-python==3.4.9.31
opencv-contrib-python-headless==4.2.0.32
opencv-python-headless==4.2.0.32
openpyxl==3.0.3
opt-einsum==3.1.0
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.7.2
parse==1.18.0
parsel==1.5.2
parso==0.5.1
path==15.0.0
pdfminer.six==20200124
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Protego==0.1.16
protobuf==3.11.3
psycopg2-binary==2.9.1
ptyprocess==0.6.0
py2app==0.21
pyasn1==0.4.7
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyDispatcher==2.0.5
pyee==8.1.0
pygame==1.9.6
Pygments==2.4.2
PyHamcrest==2.0.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyppeteer==0.2.4
pyquery==1.4.3
pyresparser==1.0.6
pyrsistent==0.15.4
pytesseract==0.3.2
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
queuelib==1.5.0
requests==2.23.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.2.0
Scrapy==2.0.0
scrapyd==1.2.1
scrapyd-client==1.1.0
scrapyrt==0.11.0
seaborn==0.10.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
soupsieve==2.0
spacy==2.2.3
spot-connect==1.0.7
srsly==1.0.2
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.3.1
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.2
Twisted==19.10.0
typing-extensions==3.7.4.3
urllib3==1.24.3
virtualenv==20.0.4
w3lib==1.21.0
wasabi==0.6.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.0
xlrd==1.2.0
XlsxWriter==1.2.8
yarl==1.6.3
zipp==3.0.0
zope.interface==4.7.1
